Better Warehouse

+++ PCPartPicker/Ram.cs
@@ -9,11 +9,11 @@ namespace PCshop
     public class Ram : Component
     {
         private int _GB;
-        private double _Mhz;
+        private int _Mhz;
         private int _ddrVersion;
 
         public int GB { get => _GB; set => _GB = value; }
-        public double Mhz { get => _Mhz; set => _Mhz = value; }
+        public int Mhz { get => _Mhz; set => _Mhz = value; }
         public int DdrVersion { get => _ddrVersion; set => _ddrVersion = value; }
 
         public Ram() : base() { }
